Add API for fetching all sizes. Add tests. Rename Size -> DropletSize.
